Move SyntaxHighlight into scala.quoted.reflect.printers
