+++ compiler/src/dotty/tools/dotc/quoted/QuoteContextImpl.scala
@@ -14,8 +14,7 @@ import dotty.tools.dotc.quoted.QuoteUtils._
 import dotty.tools.dotc.core.Decorators._
 
 import scala.quoted.QuoteContext
-import scala.quoted.show.SyntaxHighlight
-import scala.quoted.reflect.printers.{ExtractorsPrinter, SourceCodePrinter}
+import scala.quoted.reflect.printers.{ExtractorsPrinter, SourceCodePrinter, SyntaxHighlight}
 
 import scala.internal.quoted.PickledQuote
 import scala.tasty.reflect._
