+++ tests/run-macros/i6270/Macro_1.scala
@@ -1,5 +1,5 @@
 import scala.quoted._
-import scala.quoted.show.SyntaxHighlight.ANSI
+import scala.quoted.reflect.printers.SyntaxHighlight.ANSI
 
 object api {
   extension (inline x: String) inline def reflect : String =
